:recycle: Rename the Flask app to `join_github_app`
