+++ join_github_app/static/VERSION.txt
