+++ operations_engineering_landing_page.py
@@ -4,9 +4,9 @@ import logging
 from os import environ
 from dotenv import dotenv_values
 
-from landing_page_app.main.services.github_service import GithubService
-from landing_page_app.main.scripts.github_script import GithubScript
-from landing_page_app import create_app
+from join_github_app.main.services.github_service import GithubService
+from join_github_app.main.scripts.github_script import GithubScript
+from join_github_app import create_app
 
 
 def get_tokens():
