+++ tests/scripts/test_github_script.py
@@ -1,7 +1,7 @@
 import unittest
 from unittest.mock import patch, MagicMock
-from landing_page_app.main.scripts.github_script import GithubScript
-from landing_page_app.main.config.constants import (
+from join_github_app.main.scripts.github_script import GithubScript
+from join_github_app.main.config.constants import (
     MINISTRY_OF_JUSTICE,
     MOJ_ANALYTICAL_SERVICES,
     MOJ_TEST_ORG,
@@ -16,7 +16,7 @@ class TestGithubScript(unittest.TestCase):
         self.test_user = "some-user"
         self.approved_email_address = "some@justice.gov.uk"
 
-    @patch("landing_page_app.main.services.github_service")
+    @patch("join_github_app.main.services.github_service")
     def test_is_github_seat_protection_enabled_when_conditions_are_okay(
         self, mock_github_service
     ):
@@ -28,7 +28,7 @@ class TestGithubScript(unittest.TestCase):
         enabled = github_script.is_github_seat_protection_enabled()
         self.assertEqual(enabled, False)
 
-    @patch("landing_page_app.main.services.github_service")
+    @patch("join_github_app.main.services.github_service")
     def test_is_github_seat_protection_enabled_when_minimum_seats(
         self, mock_github_service
     ):
@@ -42,7 +42,7 @@ class TestGithubScript(unittest.TestCase):
         enabled = github_script.is_github_seat_protection_enabled()
         self.assertEqual(enabled, True)
 
-    @patch("landing_page_app.main.services.github_service")
+    @patch("join_github_app.main.services.github_service")
     def test_is_github_seat_protection_enabled_when_more_than_minimum_seats(
         self, mock_github_service
     ):
@@ -52,7 +52,7 @@ class TestGithubScript(unittest.TestCase):
         enabled = github_script.is_github_seat_protection_enabled()
         self.assertEqual(enabled, False)
 
-    @patch("landing_page_app.main.services.github_service")
+    @patch("join_github_app.main.services.github_service")
     def test_is_github_seat_protection_enabled_when_more_than_allowed_max_pending_invites(
         self, mock_github_service
     ):
@@ -70,7 +70,7 @@ class TestGithubScript(unittest.TestCase):
         enabled = github_script.is_github_seat_protection_enabled()
         self.assertEqual(enabled, True)
 
-    @patch("landing_page_app.main.services.github_service")
+    @patch("join_github_app.main.services.github_service")
     def test_is_github_seat_protection_enabled_when_below_max_pending_invites(
         self, mock_github_service
     ):
@@ -82,14 +82,14 @@ class TestGithubScript(unittest.TestCase):
         enabled = github_script.is_github_seat_protection_enabled()
         self.assertEqual(enabled, False)
 
-    @patch("landing_page_app.main.services.github_service")
+    @patch("join_github_app.main.services.github_service")
     def test_is_user_in_audit_log_when_no_users_in_audit_log(self, mock_github_service):
         github_script = GithubScript(mock_github_service)
         mock_github_service.get_removed_users_from_audit_log.return_value = []
         found_user = github_script.is_user_in_audit_log(self.test_user, MOJ_TEST_ORG)
         self.assertEqual(found_user, False)
 
-    @patch("landing_page_app.main.services.github_service")
+    @patch("join_github_app.main.services.github_service")
     def test_is_user_in_audit_log_when_different_users_in_audit_log(
         self, mock_github_service
     ):
@@ -101,7 +101,7 @@ class TestGithubScript(unittest.TestCase):
         found_user = github_script.is_user_in_audit_log(self.test_user, MOJ_TEST_ORG)
         self.assertEqual(found_user, False)
 
-    @patch("landing_page_app.main.services.github_service")
+    @patch("join_github_app.main.services.github_service")
     def test_is_user_in_audit_log(self, mock_github_service):
         github_script = GithubScript(mock_github_service)
         mock_github_service.get_removed_users_from_audit_log.return_value = [
@@ -110,7 +110,7 @@ class TestGithubScript(unittest.TestCase):
         found_user = github_script.is_user_in_audit_log(self.test_user, MOJ_TEST_ORG)
         self.assertEqual(found_user, True)
 
-    @patch("landing_page_app.main.services.github_service")
+    @patch("join_github_app.main.services.github_service")
     def test_add_new_user_to_github_org_with_incorrect_inputs(
         self, mock_github_service
     ):
@@ -136,7 +136,7 @@ class TestGithubScript(unittest.TestCase):
         github_script.add_new_user_to_github_org(None, [MOJ_TEST_ORG])
         github_script.github_service.get_user.assert_not_called()
 
-    @patch("landing_page_app.main.services.github_service")
+    @patch("join_github_app.main.services.github_service")
     def test_add_new_user_to_github_org(self, mock_github_service):
         github_script = GithubScript(mock_github_service)
         github_script.add_new_user_to_github_org(
@@ -144,7 +144,7 @@ class TestGithubScript(unittest.TestCase):
         )
         mock_github_service.invite_user_to_org_using_email_address.assert_called()
 
-    @patch("landing_page_app.main.services.github_service")
+    @patch("join_github_app.main.services.github_service")
     def test_get_selected_organisations(self, mock_github_service):
         github_script = GithubScript(mock_github_service)
         orgs = github_script.get_selected_organisations(None, None)
@@ -162,7 +162,7 @@ class TestGithubScript(unittest.TestCase):
         self.assertEqual(len(orgs), 1)
         self.assertEqual(orgs[0], MOJ_ANALYTICAL_SERVICES)
 
-    @patch("landing_page_app.main.services.github_service")
+    @patch("join_github_app.main.services.github_service")
     def test_validate_user_rejoining_org(self, mock_github_service):
         github_script = GithubScript(mock_github_service)
         github_script.is_user_in_audit_log = MagicMock(return_value=False)
@@ -172,7 +172,7 @@ class TestGithubScript(unittest.TestCase):
         result = github_script.validate_user_rejoining_org([MINISTRY_OF_JUSTICE], "")
         self.assertTrue(result)
 
-    @patch("landing_page_app.main.services.github_service")
+    @patch("join_github_app.main.services.github_service")
     def test_add_returning_user_to_github_org_with_incorrect_inputs(
         self, mock_github_service
     ):
@@ -193,7 +193,7 @@ class TestGithubScript(unittest.TestCase):
         github_script.add_returning_user_to_github_org(None, [MOJ_TEST_ORG])
         github_script.github_service.get_user.assert_not_called()
 
-    @patch("landing_page_app.main.services.github_service")
+    @patch("join_github_app.main.services.github_service")
     def test_add_returning_user_to_github_org(self, mock_github_service):
         github_script = GithubScript(mock_github_service)
         mock_github_service.github_service.get_user.return_value = "a real user"
@@ -202,7 +202,7 @@ class TestGithubScript(unittest.TestCase):
         )
         mock_github_service.invite_user_to_org_using_nameduser.assert_called()
 
-    @patch("landing_page_app.main.services.github_service")
+    @patch("join_github_app.main.services.github_service")
     def test_add_returning_user_to_github_org_with_incorrect_org_name(
         self, mock_github_service
     ):
